chore: Remove deprecated config.ts file and update common.ts with new TOML parsing library
The config.ts file has been deleted as it is no longer needed. The TOML parsing library used in the common.ts file has been updated to '@ltd/j-toml' for compatibility with Cloudflare Workers.
