+++ src/common.ts
@@ -1,4 +1,5 @@
 import { RouterHandler } from "@tsndr/cloudflare-worker-router";
+import * as TOML from '@ltd/j-toml';
 
 export type Env = {
     GITHUB_APP_ID: string;
@@ -23,6 +24,24 @@ export type Config = {
 
     checkPermissions(user: string): boolean;
 };
+
+const parseTomlConfig =
+    async (uri: string): Promise<Config | undefined> => {
+        const response = await fetch(uri);
+        const parsedToml = TOML.parse(await response.text());
+        // TODO: This would need to be addressed in the new version of the 
+        // permissions config file.
+        const admin = parsedToml.admin;
+        if (!admin || !admin.name) {
+            throw new Error('Error while parsing the config file, config file is invalid');
+        }
+        return {
+            admin: { name: admin.name },
+            checkPermissions: (user: string) => admin.name === user,
+        };
+    };
+
+export { parseTomlConfig };
 
 export const sleep = (ms: number) => {
     return new Promise(resolve => setTimeout(resolve, ms));
